# Conflicts:
#	go.mod
#	go.sum

+++ clients/hadoopfs/src/main/java/io/lakefs/LinkOnCloseOutputStream.java
@@ -18,6 +18,7 @@ class LinkOnCloseOutputStream extends OutputStream {
     private final URI physicalUri;
     private final MetadataClient metadataClient;
     private final OutputStream out;
+    private boolean isLinked;
 
     /**
      * @param lfs LakeFS file system
@@ -63,11 +64,14 @@ class LinkOnCloseOutputStream extends OutputStream {
 
         // Now the object is on the underlying store, find its parameters (sadly lost by
         // the underlying Hadoop FileSystem) so we can link it on lakeFS.
-        try {
-            lfs.linkPhysicalAddress(objectLoc, stagingLoc, physicalUri, metadataClient);
-        } catch (io.lakefs.clients.api.ApiException e) {
-            throw new IOException("link lakeFS path to physical address", e);
+        if (!this.isLinked) {
+            try {
+                lfs.linkPhysicalAddress(objectLoc, stagingLoc, physicalUri, metadataClient);
+            } catch (io.lakefs.clients.api.ApiException e) {
+                throw new IOException("link lakeFS path to physical address", e);
+            }
+            this.isLinked = true;
+            lfs.deleteEmptyDirectoryMarkers(new Path(objectLoc.toString()).getParent());
         }
-        lfs.deleteEmptyDirectoryMarkers(new Path(objectLoc.toString()).getParent());
     }
 }
